refactor(port): typed Confidence and Score — domain types for metrics (TSK-284)
Confidence (float64, 0.0-1.0) for pattern detection confidence.
Score (float64, 0-100) for SOLID/Hexa/SymbolQuality compliance scores.
Updated struct fields in clinic, protocol. Explicit casts at computation
sites document where raw floats become domain values.

+++ internal/clinic/hexa_test.go
@@ -201,7 +201,7 @@ func TestComputeHexaViolations_MultipleRules(t *testing.T) {
 
 	// Score: (4 - 3) / 4 * 100 = 25%
 	expectedScore := 25.0
-	if report.Score != expectedScore {
+	if float64(report.Score) != expectedScore {
 		t.Errorf("expected score %.1f, got %.1f", expectedScore, report.Score)
 	}
 }
